pyproject.toml
[project]
name = "catxgnn"
version = "0.1.0"
description = "Canada drug repurposing predictions using TxGNN knowledge graph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
    "lxml>=4.9.0",
    "tqdm>=4.65.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/catxgnn"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
